requirements.txt
pandas
numpy
transformers
torch
pyyaml
langgraph
flask
flask-socketio